Move grading logic into own module
